Moved code into aucgtk to execute after mirrorlist dialog
